Write out the javascript files in binary mode to prevent any futzing-with by the OS or Ruby
This is kind of a stab in the dark, I'm seeing files generated slightly
differently on different machines and I'm not entirely certain what's causing
it. My current theory is that there is some system, environment or other
setting that is resulting in "text mode" files having a newline appended
automatically to them.